+++ posthog/CHANGELOG.md
@@ -1,8 +1,7 @@
 ## Next
-- Allow collecting FCM device token in SDK ([#376](https://github.com/PostHog/posthog-android/pull/376))
 
 ### Added
-- Allow collecting FCM device token in SDK core ([#396](https://github.com/PostHog/posthog-android/pull/396))
+- Allow collecting FCM device token in SDK ([#376](https://github.com/PostHog/posthog-android/pull/376))
 
 - feat: Add `getFeatureFlagResult` to `PostHogFeatureFlagsInterface`, drop `getFeatureFlag`, `getFeatureFlagPayload`. ([#398](https://github.com/PostHog/posthog-android/pull/398))